improve: Change the mock-data to JSX elements and mix them
So I tried to make a method that changes the state value(selectedIndex) as an onClick property
when creating the DessertItem component through mock-data. For this, I used "useEffect" but i
meet one problem that called "missing dependency". In "useEffect", the method that change the state
value is given as the prop of DessertItem, I think it seems not be the way. So now I`m thinking about how to solve this problem